Add files via upload
The objective of this project is to develop a machine learning model capable of classifying images of flowers into their respective species based on visual features. The dataset comprises 210 images, each sized at 128x128 pixels, across 10 different species of flowering plants. The images are in .png format, and the species labels are provided as integers in a separate file, flower-labels.csv. *Only includes half of the image dataset, as github doesn't allow for uploading >100 files at once.*